Moved code from PrettyChartDesignViewModel.cs to design itself reduciong file pollution in project and localizing code for prettycharts
