+++ UI/Charts/PrettyChartDesign.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using GolemUI.Model;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -12,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using static GolemUI.Model.PrettyChartData;
 
 namespace GolemUI.UI.Charts
 {
@@ -24,5 +26,44 @@ namespace GolemUI.UI.Charts
         {
             InitializeComponent();
         }
+    }
+
+    class PrettyChartDesignViewModel
+    {
+        public int NumberOfBins => ChartData.BinData.BinEntries.Count;
+
+        public PrettyChartData ChartData => new PrettyChartData()
+        {
+            NoAnimate = true,
+            BinData = new PrettyChartBinData()
+            {
+                BinEntries = new List<PrettyChartBinEntry>() {
+                    new PrettyChartBinEntry(){Label="Piąta", Value=5.5},
+                    new PrettyChartBinEntry(){Label="Szósta", Value=2.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Piąta", Value=5.5},
+                    new PrettyChartBinEntry(){Label="Szósta", Value=2.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Piąta", Value=5.5},
+                    new PrettyChartBinEntry(){Label="Szósta", Value=2.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Piąta", Value=5.5},
+                    new PrettyChartBinEntry(){Label="Szósta", Value=2.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Piąta", Value=5.5},
+                    new PrettyChartBinEntry(){Label="Szósta", Value=2.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                    new PrettyChartBinEntry(){Label="Siódma", Value=6.5},
+                }
+            }
+        };
     }
 }
